Need help for find error !!!!!!

+++ Weather/app/src/main/java/AppAndroidKotlin/weather/weather/Weather.kt
@@ -2,6 +2,6 @@ package AppAndroidKotlin.weather.weather
 
 data class Weather(val description: String,
                    val temperature: Float,
-                   val humidity: Float,
+                   val humidity: Int,
                    val pressure: Int,
                    val iconUrl: String)
